refactor: move exec to myexec
