+++ hrp/cmd/pytest.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/httprunner/httprunner/v4/hrp/internal/builtin"
+	"github.com/httprunner/httprunner/v4/hrp/internal/myexec"
 	"github.com/httprunner/httprunner/v4/hrp/internal/pytest"
 	"github.com/httprunner/httprunner/v4/hrp/internal/version"
 )
@@ -23,7 +23,7 @@ var pytestCmd = &cobra.Command{
 		packages := []string{
 			fmt.Sprintf("httprunner==%s", version.VERSION),
 		}
-		_, err := builtin.EnsurePython3Venv(venv, packages...)
+		_, err := myexec.EnsurePython3Venv(venv, packages...)
 		if err != nil {
 			log.Error().Err(err).Msg("python3 venv is not ready")
 			return err
