+++ hrp/internal/scaffold/main.go
@@ -4,7 +4,6 @@ import (
 	"embed"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"time"
 
@@ -14,6 +13,7 @@ import (
 
 	"github.com/httprunner/httprunner/v4/hrp"
 	"github.com/httprunner/httprunner/v4/hrp/internal/builtin"
+	"github.com/httprunner/httprunner/v4/hrp/internal/myexec"
 	"github.com/httprunner/httprunner/v4/hrp/internal/sdk"
 	"github.com/httprunner/httprunner/v4/hrp/internal/version"
 )
@@ -177,7 +177,7 @@ func CreateScaffold(projectName string, pluginType PluginType, venv string, forc
 func createGoPlugin(projectName string) error {
 	log.Info().Msg("start to create hashicorp go plugin")
 	// check go sdk
-	if err := builtin.ExecCommandInDir(exec.Command("go", "version"), projectName); err != nil {
+	if err := myexec.RunCommand("go", "version"); err != nil {
 		return errors.Wrap(err, "go sdk not installed")
 	}
 
@@ -209,7 +209,7 @@ func createPythonPlugin(projectName, venv string) error {
 		fmt.Sprintf("funppy==%s", fungo.Version),
 		fmt.Sprintf("httprunner==%s", version.VERSION),
 	}
-	_, err = builtin.EnsurePython3Venv(venv, packages...)
+	_, err = myexec.EnsurePython3Venv(venv, packages...)
 	if err != nil {
 		return err
 	}
