+++ hrp/internal/wiki/main.go
@@ -1,10 +1,9 @@
 package wiki
 
 import (
-	"os/exec"
-
 	"github.com/rs/zerolog/log"
 
+	"github.com/httprunner/httprunner/v4/hrp/internal/myexec"
 	"github.com/httprunner/httprunner/v4/hrp/internal/sdk"
 )
 
@@ -14,5 +13,5 @@ func OpenWiki() error {
 		Action:   "hrp wiki",
 	})
 	log.Info().Msgf("%s https://httprunner.com", openCmd)
-	return exec.Command(openCmd, "https://httprunner.com").Run()
+	return myexec.RunCommand(openCmd, "https://httprunner.com")
 }
